Refs 320, organize customer user forms
