+++ hymaintenance/high_ui/views/users.py
@@ -9,18 +9,18 @@ from django.views.generic import FormView
 from django.views.generic import TemplateView
 from django.views.generic import UpdateView
 
-from customers.forms import AdminUserCreateForm
-from customers.forms import MaintenanceUserModelForm
-from customers.forms import MaintenanceUserProfileUpdateForm
-from customers.forms import ManagerUserCreateForm
-from customers.forms import ManagerUsersUpdateForm
-from customers.forms import OperatorUserArchiveForm
-from customers.forms import OperatorUserCreateForm
-from customers.forms import OperatorUserCreateFormWithCompany
-from customers.forms import OperatorUsersUpdateForm
-from customers.forms import OperatorUserUnarchiveForm
-from customers.forms import StaffUserProfileUpdateForm
-from customers.forms import StaffUserUpdateForm
+from customers.forms.users.create_user import AdminUserCreateForm
+from customers.forms.users.create_user import ManagerUserCreateForm
+from customers.forms.users.create_user import OperatorUserCreateForm
+from customers.forms.users.create_user import OperatorUserCreateFormWithCompany
+from customers.forms.users.user_base import MaintenanceUserModelForm
+from customers.forms.users.user_base import StaffUserUpdateForm
+from customers.forms.users.user_profile import MaintenanceUserProfileUpdateForm
+from customers.forms.users.user_profile import StaffUserProfileUpdateForm
+from customers.forms.users_list.list import OperatorUserArchiveForm
+from customers.forms.users_list.list import OperatorUserUnarchiveForm
+from customers.forms.users_list.list_by_company import ManagerUsersUpdateForm
+from customers.forms.users_list.list_by_company import OperatorUsersUpdateForm
 from customers.models.user import MaintenanceUser
 from maintenance.forms.consumer import MaintenanceConsumerModelForm
 from maintenance.forms.consumer import MaintenanceConsumersUpdateForm
